+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2018-05-29
+### Added
+- Moderator management page
+- Moderator notification management page
+
+### Fixed
+- Prevent whitespace under banners
+
 ## [0.5.1] - 2018-04-24
 ### Changed
 - Back to pinned osf-style commit